Add ability to generate a CHANGELOG.md
Added the [github-changelog-generator](https://github.com/skywinder/github-changelog-generator) gem to the project so that once we have linked the project to a GitHub repo we can automatically generate and maintain a CHANGELOG.md file for the project.

github-changelog-generator generates the log based on the tags, issues, labels and pull requests it finds in GitHub for the repo.

To simplify its use we have also added a rake task you can call (`bundle exec rake changelog`) to update CHANGELOG.md.