Repurpose project as WCR back office
We've decided to go a different way with the entity matching service. Rather than building a whole new application in rails we're instead going to migrate the existing Java code into the Java Waste carriers service, and whilst doing that strip the dependency on Elasticsearch.

We still need to build a new back office though, and as we haven't implemented any specific entity matching logic yet it means this can be re-purposed as the base code project for it.

Hence this change simply goes in and replaces any reference to entity matching with wcr back office, and updates its config to be able to connect to the standard registrations and users databases.