Set up authentication using Devise (#4)
* Set up authentication using Devise

https://eaflood.atlassian.net/browse/WC-126

Only valid, authenticated users should be able to access the app. We use Devise to authenticate.

* Install Devise

* Set up views and config to match WEX and FRAE

Our back offices should be consistent. This commit sets up the views and updates the configuration to match what we used for other apps. This means installing SimpleForm for form generation, and adding the Invitable extension for Devise.

Users are now stored in the back_office_users collection.

* Test user password validation

This includes setting up FactoryBot for creating test users.

* Delete unnecessary ApplicationHelper

We only needed this to define a current_user method. Devise now handles this for us, so we can delete it (and get 100% test coverage).

* Only logged-in users can access the app

Also set up database cleaning for tests.