Mount renewals journey in back office (#22)
* Mount renewals journey in back office

https://eaflood.atlassian.net/browse/WC-123

Users should be able to renew via an assisted digital route. The back office should support this.

Now that we have the WasteCarriersEngine, we can mount it in the back office to get the renewals journey.

Once we have the journey working in the app, we can start making any additional changes necessary for the AD process.

* Install the engine, rip out Devise

Mount the WasteCarriersEngine and add it to the routes.

We already have some Devise config in the engine, so it's easier to tear it out and start over.

* Use AD Worldpay test config

We have a different test environment for AD Worldpay payments, so let's make sure we use the right config.

* Add engine initialiser

This is so we can include the engine version on the version page.

* Load engine from master branch

We've merged the engine setup into master of waste-carriers-renewals now.

* Rip out user tests

The engine is currently handling all the user behaviour, so those tests belong in the engine repo, not here.