Move users out of engine (#46)
* Move users out of engine

When we first set up the engine, we included users and related functionality like Devise and CanCanCan. However, it's become clear that the front office and back office users are sufficiently different that we shouldn't be trying to tackle this in the engine and should let the host apps handle it instead.

This PR re-implements the user functionality (including Devise and CanCanCan) which was removed from the engine in DEFRA/waste-carriers-renewals#228

* Copy over Devise and CanCanCan bits

Add all the config, models and tests that were removed from the engine.

Temporarily using the feature/move-users-out-of-engine branch of the engine, as otherwise none of it will work.

* Fix module name and inconsistencies with front-office

The application.rb config was using an incorrectly-named module. This appeared to be causing some oddities with environment variables.

I also made a number of small tweaks to make it more consistent with the waste-carriers-front-office.

* Include ApplicationHelper from engine

We were getting intermittent undefined method errors for methods in the WasteCarriersEngine::ApplicationHelper. Adding this line seems to fix the issue.