Mount renewals engine in a directory, not root (#48)
* Mount renewals engine in a directory, not root

https://eaflood.atlassian.net/browse/WC-427

Currently the waste-carriers-back-office mounts the renewals engine at root. (For example, /renew/CBDU1234).

The new and existing applications will be accessed through the same domain. Determining which requests should go to which applications. This is a lot easier to do if the routes are similar. Mounting the journey within a directory (eg /fo/renew/CBDU1234) means we can cover the whole engine with a single line (eg /fo/*) rather than specifying every single route, which would be time-consuming to set up and maintain.

This change should cover the Devise paths as well.

* Mount engine and Devise at /bo

Stands for back office.