Switch from govuk_admin_template to govuk_template (#67)
* Switch from govuk_admin_template to govuk_template

The admin template has now been deprecated and the new admin layout is currently marked as 'experimental' in the GOV.UK component guide, meaning breaking changes are likely: https://govuk-publishing-components.herokuapp.com/component-guide/layout_for_admin

We have also encountered a bizarre bug with the error message links, which seems to be fixed by removing the admin template (see https://eaflood.atlassian.net/browse/WC-436)

For these reasons, we've decided to switch templates and use govuk_template for now. This also means we can have greater consistency across both office applications, and speed up the migration to the new design system when the time comes.

* Remove govuk_admin_template

* Install govuk_template

We also include some of the partials we use in the front-office.

There are a few differences - we don't include the beta bar, the service name in the header is different, and we don't have additional text in the footer.