List in-progress registrations (#76)
* List in-progress registrations

https://eaflood.atlassian.net/browse/WC-374

Logged-in users should be able to see a list of transient_registrations, aka in-progress renewals. Each listed renewal should include a link to further details.

This is the first part of a larger story. The next step will be adding a search and pagination.

* Restrict dashboard to signed-in users

* Display transient registrations

And include a link to the details page.

More work is required to decide what to display for each registration and how, but this does at least link up the back office journey.

* Move text to I18n