Fix routing to ensure /bo/ path is included (#75)
https://eaflood.atlassian.net/browse/WC-440

When we dev locally we don't have to think too much about how requests are routed because we call apps directly using their address and port number.

When we go to production though, this is not the case and all access is via 2 urls (examples)

- https://waste-carriers-dev.aws.defra.cloud
- https://admin-waste-carriers-dev.aws.defra.cloud

The problem is each of these external url's is handling requests for 2 apps, so the web server that handles the incoming requests needs to know how to route traffic to the relevant app based on what comes in.

We agreed the routing logic could be based on the path (the bit of the URL that appears after the domain above). So admin-waste-carriers-dev.aws.defra.cloud will go to the frontend app, whereas admin-waste-carriers-dev.aws.defra.cloud/bo/ will go to the back office app.

The problem we currently have in the back-office app is that after certain actions you are redirected to the app root, for example sign in. Because of this when we run it in a production like environment, the redirect either takes you to the other system, or causes the app to error.

This job is about fixing the issue in the back office so that all redirections return a user to a url that includes the /bo/ path.

With these changes made in this commit when a user comes to the app e.g. localhost:8000 they'll be redirected to `/bo`. However if not authenticated they'll then be redirected to the user sign in page.

Once authenticated they'll be redirected to `/bo` rather than just `/` as we know that'll just redirect folks to `/bo` anyway.

Finally when users sign out they'll be redirected to `/bo/users/sign_in` rather than `/`, because again they'll eventually land there anyway.

The approach to doing this was taken from https://github.com/plataformatec/devise/wiki/How-To:-redirect-to-a-specific-page-on-successful-sign_in,-sign_out,-and-or-sign_up. There is an alternative approach https://github.com/plataformatec/devise/wiki/How-To%3A-Redirect-to-a-specific-page-on-successful-sign-up-(registration) which moves the redirects to a new controller, but as we don't expect to add anything more to `application_controller.rb` (famous last words!) it was not felt necessary at this time.

The main other change was the ordering of routes in routes.rb

Based on our knowledge of rails the ordering of routes matters as this sets the order of priority in the rails router.

When trying out some possible solutions to this issue, we found we could not get to some of the routes set because they instead were being picked up by the WCR engine. Moving the mount call to the end of routes.rb resolved the issue.

So hopefully the new ordering of the routes means

- calls to `/` get picked up first (and redirected)
- specific calls to `/bo/#` get picked up by the app
- any other calls to `/bo/#` will be passed to the engine