Implement search for in-progress renewals (#80)
* Implement search for in-progress renewals

https://eaflood.atlassian.net/browse/WC-374

This PR will implement searching for in-progress renewals from the back office dashboard.

Users should be able to search by registration number, company name, contact name or postcode.

It should also be possible to filter the search by what actions are required (eg. renewal still in progress, pending payment, pending conviction check).

* Implement search by reg_identifier

Adding `?term=foo` to the URL for the GET request allows you to search for transient_registrations with a matching reg_identifier.

Further work is needed to look into case sensitivity, sanitising the search term, searching against multiple attributes, etc.

* Add search field to page

Also list the attributes you can search by. Only one of these is implemented so far but we'll get there!

I also removed the button styling from the 'details' links as GDS guidance is to only have one button per page. The search button feels more important.

* Include company_name in searches

We now display a transient_registration if there is an exact match on reg_identifier or company_name.

* Fix accidentally pointless tests

Our specs for the search results were checking to see if the term appeared on the page to determine whether the correct result was shown. However, the search term will always appear in the search field, regardless of whether there are matches or not. This meant the tests were completely missing a search-braeking bug.

Checking for the presence of the link to the details page for that renewal is a more precise way to check the results.

* Search for matching postcodes

Users can now search for a matching renewal by postcode.

* Search by last name of contact person

The search in waste-carriers-frontend says you can search by "contact name". However what it's actually doing is searching by the lastName field. So I matched that for consistency, but updated the content to be clearer.

* Search using a regex with the term

This allows us to do case-insensitive matches, or match when the term is just part of the matching value (for example, a search for "Foo Co" should return a company called "Foo Co Ltd.")

* Match reg_identifiers exactly

Searching for "CBDU1" and getting "CBDU12", "CBDU124453" etc would be annoying. So we will use an alternative regex to be more precise for this attribute.

* Use new transient_registration methods for statuses

As of https://github.com/DEFRA/waste-carriers-renewals/pull/245, the transient_registrations have some new methods to help determine their current status and any pending actions. This means we can remove a now-duplicated method from the back office and clean up some of the other conditionals.

* Add non-functional filters to search view

These three checkboxes will be used for filtering. We save the params they pass in and then pre-tick the relevant checkboxes if that filter is 'active'. Filtering doesn't do anything yet...

* Implement filters

Use the filter params to select which transient_registrations we want to display. This stacks with the search term, so only renewals which are relevant to both should be displayed.

Also lots of refactoring.

* Refactor filter and search tests

A lot of these are making the same request, so we can reduce duplication.

* Style up the search and filters

Make the filters look a bit nicer, and try to make it clearer that they interact with the search.

* Add extra explanation of filters

Attempting to make it clearer that you can use the filters independently of the search.

* Add a 'clear all' link

This just links back to /bo, but without any filters or search in place. This should make it quicker for users to reset to the default.

* Rename method as suggested in feedback

* Move dashboard searching and filtering into a service

As discussed in the review. This makes the controller a lot lighter.

This also led to some refactoring of the tests, which cut down on the length a lot.

* Set filters separately instead of as array

This simplifies things a bit more. If we end up having more filters in the future, we may return to the array for tidiness.

* Move more logic out of index method

This was still looking a little bulky, so let's move some of the clutter into separate methods to keep it nice and clean.

* Remove deprecated static attributes from factories

Most of this work was done in https://github.com/DEFRA/waste-carriers-back-office/pull/87, but there were a few changes made in this branch which weren't covered.