Back office dashboard QA fixes (#90)
* Back office dashboard QA fixes

https://eaflood.atlassian.net/browse/WC-374

Making several fixes following QA feedback:

- clarify that postcode may not be present yet
- make sure times are displayed in current UK time zone
- move “no results found” to more readable location

* Clarify that postcodes may not always be searchable

We ask users to reconfirm their postcode during the renewal process. If a user hasn't gotten to that stage yet, there won't be a postcode to search. So let's clarify that in the microcopy.

* Display 'no results found' more obviously

We previously relied on the pagination to display this message. However, as it floats to the right, it's not very clearly visible.

Instead, let's display it left-aligned, and clearly separated from the search fields by a divider.

* Display times in UK time zone

When displaying times (as in last_modified), use the current UK timezone (GMT or BST).