Add page showing conviction match details (#91)
* Add page showing conviction match details

https://eaflood.atlassian.net/browse/WC-376

This page should display whether the user has declared convictions, and also the details of any conviction matches. The back office users should be able to access it from the transient registration details page.

* Create blank convictions page

Each transient registration should have a convictions page for reviewing convictions info. This sits at /bo/transient-registrations/CBDU1234/convictions.

If there are no declared convictions or matches, we may not need to link to this, but it still can be loaded.

* Display overview of convictions info, link between pages

There are three types of conviction info - declared_convictions, matches for the business and matches for key people. The page now displays a single-sentence overview of the status for each, plus a line at the top stating whether a check is required or not.

The convictions page now includes a 'back' link to the transient registration details, and the conviction check buttons on the details page now link to the convictions page. It also made sense to add a 'back' link from the details page to the dashboard so things are more joined-up and consistent.

* Display conviction match data

If the match is positive or unknown, display data from the conviction_search_result.

* Display relevant people without matching convictions

If the user declared that someone had a relevant conviction, but there were no matches for them, we want to know about it.

This commit also adds some details about each person to aid with making the checks. Since we're reusing some of the rows we use in the transient registration page, let's refactor to use a partial instead.

* Improve the words

Trying to make the language a little more natural.

* Add accept and reject buttons

They don't do anything yet – that's for a future PR.

Also add some styling and fix some indentation.