Users can approve or reject based on convictions (#93)
* Users can approve or reject based on convictions

https://eaflood.atlassian.net/browse/WC-376

After reviewing convictions, a back office user should be able to approve or reject the renewal, and leave a comment explaining the decision.

Approving won't cause the renewal yet (see WC-302) but should update the conviction_sign_off.

Rejecting should change the status of the renewal, but leave the original registration alone.

* Set up routes, controllers and views for approvals and rejections

Approval and rejection pages should be located at /bo/transient-registrations/CBDU1234/convictions/approve or /reject.

* Add some content to pages, link from conv details

The buttons on the convictions details page should link to the approve and reject pages.

* Working forms for accepting or rejecting conviction

These are workin in the sense that submitting them updates the metaData.revoked_reason. The actual approval or rejection doesn't happen yet.

For these forms, I went back to using form objects and a simplified FormsController to manage loading and submitting the forms. I inherited from the engine where possible, but especially for the controller, there's a lot of functionality (eg managing the workflow state) which means inheritence isn't feasible.

* Update the conviction_sign_off when approving

When a back office user approves a renewal during a conviction check, update the conviction_sign_off to record who approved it and when.

* Revoke transient_registrations when rejecting for convictions

This only affects the renewal, not the existing registration.

* Use engine to update conviction_sign_off

We can use the new method added in https://github.com/DEFRA/waste-carriers-renewals/pull/252 to simplify our code.

* Display results of conviction sign off on transient reg page

Display any information about the conviction_sign_off, as well as the reason. Also update the summary message for convictions.

* Move conviction sign off info to convictions page

But link to the convictions page, and keep the summary message.

* Can't approve or reject convictions which are already signed off

Throw a validation error if this is attempted.