Set up roles for back office users (#96)
* Set up roles for back office users

https://eaflood.atlassian.net/browse/WC-377

There are a number of different roles in the existing system:

- agency user
- agency user with refund
- finance user
- finance admin
- agency admin
- finance super

We need to replicate these in the new back office so we can assign permissions to them.

* Seed different types of users

Also modify the seed script to only run if the environment is not production or we specify a particular env var.

They currently all have the same role – this will change.

* Define roles

List the valid roles. Each user has a role stored as a string against their record. Roles must be on the list to be valid.

* Define actions and who can do them

Use CanCanCan to define the actions which require specific permissions and define which users can do them.

The next step is to actually implement these checks.

* Only agency users can approve/reject convictions

All back office users can see conviction info, but only agency and agency_with_refund users can approve or reject a renewal based on this info.

This also acts as a sort of proof-of-concept for other future authorising, eg. for payments.

* Only agency users can access AD journey

The forms mounted for the engine are now locked down to agency and agency_with_refund users only. Finance users should not be able to deal with the non-finance aspects of a renewal.

* Remove unnecessary actions

Because `can :update` covers everything in the AD journey, we can simplify the different permission actions.

Also tweak the factories slightly as we've merged in changes from master which cover this.

* Hide link to AD for non-agency users

Also add an explanatory comment to `can :update`.