Back office users can select which type of payment they want to add (#101)
* Back office users can select which type of payment they want to add

https://eaflood.atlassian.net/browse/WC-377

There are 5 different types of payment which can be manually added. The type of payment a user can add also depends on their role.

Rather than sending the user straight into a form, we first ask them to select which type of payment they want to record, and then direct them to the correct form for that type.

* Set up payment type form

Set up controller, view and route for the payment form. This lets users select which type of payment they want to make.

Submitting the form doesn't do anything yet.

* Start working on routing based on form answer

We don't have anything to route to yet (waiting on https://github.com/DEFRA/waste-carriers-back-office/pull/97) but we can implement what happens if the form doesn't submit a valid payment method.

Storing a single list of valid payment methods also means we can use this to reduce duplication in the form.

* Disable payment options not permitted to user

If the user can't make a certain payment option, disable the relevant form field and let them know it's unavailable.

This also makes clear that it is possible to record a payment of that type – just not with this type of account.

* Fix authorisation for transfer payment form

This ability was renamed before this form was merged in, so the tests were broken.

* Display payment type and method in a more readable way

Make sure we display it in the correct case with spaces between words, rather than dumping the DB value onto the page.

* Redirect to correct payment form after submission

When the form is submitted with a valid payment_type, find the correct path for that payment type form and redirect to it.

Currently only the transfer form exists - others will be added shortly in separate PRs.

* Link from transient reg page to payment page