Renewals are completed once there are no pending issues (#108)
* Renewals are completed once there are no pending issues

https://eaflood.atlassian.net/browse/WC-302

Renewals will not be automatically completed if there is a pending conviction check or payment has not been recorded. However, once the conviction check has been approved and the balance is zero, the renewal should automatically complete. This should happen once the last pending action is cleared.

Back office users should also be made aware that clearing a pending issue (eg recording a payment which brings the balance to zero) will also complete the renewal.

* Create RenewabilityCheckService

This new service can be used to check if a renewal is ready to complete (eg there are no pending checks) and to trigger the completion of the renewal (if valid).

* Renew after conviction approval form submitted

Once the conviction approval form is successfully submitted, attempt to renew if there are no more pending checks.

A lot of our tests relied on checking the content of the transient_registration after submission. If the renewal completes, that transient_registration no longer exists, so we block completion with a stub for most tests.

* Renew if possible after recording cash payment

If all conditions are met, automatically renew after saving the payment.

* Renew if possible after saving cheque payment

* Add env vars to Travis

Hoping this will fix some test failures which aren't happening locally.

* Renew if possible after saving postal order payment

* Renew if possible after saving transfer payment

* Renew if possible after saving Worldpay payment

* Redirect to dashboard if no matching transient reg exists

Normally, after updating a payment or conviction check, the app would redirect to the transient_registration page. However, if the renewal has completed, the transient_registration won't exist and so neither will this page.

Instead, if someone tries to view a non-existent transient_registration, just redirect them to the dashboard.

* Add warnings to payment and conv pages about renewal

We need to make it clear that performing certain actions will cause the renewal to complete immediately.

* Don't link to further actions if status REVOKED

If an application is rejected during a conv check, the transient reg page should say this and not link to actions which would normally complete it.

* Display 'rejected' tags on dashboard

If an application is rejected, display the appropriate tag as its status when it's listed on the dashboard.

* Convert amount to pence only when creating payment

This fixes a bug discovered during testing. If a payment is added but triggers a validation error, the value in the amount field had already been multiplied by 100. This can lead to an accidental stacking effect when a valid form is eventually submitted with a significantly inflated value. Only converting the value to pence before we pass it in to the method which creates the payment avoids this problem.

* Hide accept/reject on conv check page when rejected

After a conviction check takes place, the conviction details page can still be accessed. However, we should not display links to approve or reject it any more.