NCCC users can send renewal stuck in WorldPay back to payment summary (#136)
* NCCC users can send renewal stuck in WorldPay back to payment summary

https://eaflood.atlassian.net/browse/WC-487

On the transient registration overview page in the back office, if the current state is worldpay_form, the "Continue renewal" button should be replaced with a "Send back to payment summary" button.

When clicked, this button should change the workflow_state from worldpay_form to payment_summary_form and then redirect the user to the payment summary form.

This should allow NCCC to help out users who have somehow gotten stuck in the WorldPay stage.

* Create revert-to-payment-summary route

Currently this just redirects to the renewal journey.

* Update state and redirect to payment summary

If the workflow_state is currently worldpay_form, change it to payment_summary_form and then redirect to that form.

* Create revert_to_payment_summary ability

This should be limited to users who can go through the renewals process (eg agency users).

* Check user permissions when reverting to payment summary

Implement the new ability check in the controller.

* Add button to transient_registration overview

If the workflow_state is WorldPay, display a message saying so. If the user can revert the renewal back to payment summary, display an additional message and a button to do it.

* Persist change to workflow_state

Because the state change wasn't saved, attempting to go back to the payment summary form just redirected to WorldPay, as that was still the saved state.

Fixed this, and added another test to avoid it happening again.

* Log and report when a renewal is reverted

So we can make sure this feature is working correctly, we log its usage and send a report to Errbit as well. We may decrease this in the future once we've confirmed it is effective.