Don't automatically list all renewals in back office dashboard (#140)
* Don't automatically list all renewals in back office dashboard

https://eaflood.atlassian.net/browse/WC-485

Currently the back office dashboard displays all renewals by default. The issue is now we have a number in progress, it makes coming back to this page quite slow.

Ideally the page would only list results when a user searches or filters. This should reduce the page load time.

* Don't get transient_registrations if no search or filters specified

Amend TransientRegistrationFinderService to return an empty array if no filters or search terms are provided.

* Don't display 'no results found' unless searched or filtered

If the user hasn't specified a search term or a filter, don't display 'no results found', as we won't have tried to find any results in the first place.