Add rake task to fix addresses in incorrect format (#143)
* Add rake task to fix addresses in incorrect format

https://eaflood.atlassian.net/browse/WC-500

Once we roll out our fix to make waste-carriers-renewals and waste-carriers-frontend save addresses from OS Places in the same format, we will still have some addresses which have already been saved.

Running this rake task should find the addresses and correct them, so all addresses are in a consistent format.

* Add working rake task to update addresses

This can be run with `bundle exec rake address:update_from_os_places`. It will only touch registrations which have a past_registration, and transient_registrations which have addresses. And then it will only attempt to update addresses which were originally generated through OS Places results. This means we can use the UPRN to retrieve the data that was previously selected by the user, but this time populate the fields in the new, more consistent way.

* Request individual address details with UPRN

Rather than getting a list using a postcode and then filtering, we can simplify the process and the amount of data we're handling.