Improve address update rake task (#150)
* Improve address update rake task

https://eaflood.atlassian.net/browse/WC-509

This task needs better logging so we can test it more thoroughly.

While adding the logging, we also discovered some bugs which needed to be addressed.

* Create new logging method

If the address has changed, output the old and new versions.

* Fix Registration query

`past_registrations` is plural, so a registration will never just have `past_registration`. This meant we weren't capturing any registrations in the task.

* Preserve first_name and last_name on postal address

This isn't added by the OS Places check but is done when we finish renewing the registration. So we need to make sure we don't override it.

* Only persist new address if it's different

Rather than hammering our database by saving addresses which haven't changed, let's just save the address if we know it is different.

This also means we don't output messages about unchanged (transient_)registrations to the log.