Use paging in fix dupe house no. rake task (#152)
https://eaflood.atlassian.net/browse/WC-509

When running the rake task in our pre-prod environment we kept getting timeouts after 2 minutes. When we've encountered issues like this before we've modified the task to use paging, with the intention that it keeps the connection to MongoDb alive because it continues to fire a query request every x registrations its processed.

However in this case it turns out that the timeout was actually a Jenkins config setting issue. However having made the changes to use paging, we still think there is value in using this method to ensure this fix does not effect the live service when being run, by trying to hold too much data in memory.

Also as a final protection against possible timeouts, we have split the task into 2 separate rake tasks, that can then be called separately in Jenkins.