Transfer a registration to an existing account (#153)
* Transfer a registration to an existing account

https://eaflood.atlassian.net/browse/WC-61

Sometimes NCCC will need to transfer ownership of a registration to a different account.

As a first step, we will set up transfer of a registration to an account which already exists.

* Set up RegistrationTransferService

This new service will deal with transferring registrations from one account to another.

* Create ExternalUser model

The back-office now needs to know that front-office users exist, so we create a new model to deal with them.

To minimise duplication, we can store most of our Devise config in a concern, which we will include in both the User and ExternalUser models.

* Swap out duplicated Devise config for concern

The new CanBehaveLikeUser concern has most of our Devise config, so we can now include that in the User model and strip outa lot of duplication.

* Find the external user to transfer the registration to

When we call transfer_to_user on a RegistrationTransferService and pass in an email, we try to find an external user with that matching email and set it to an instance variable.

If there is no match, for now we set it to nil.

* Return if no matching user exists

Since we aren't handling creation of new users yet, we should stop the transfer process when we realise there is no match.

* Update account_email values and save

Update the account_email values for the registration and transient_registration, if one exists.

* Add mailer for sending transfer emails

This sends a basic email to a registration's account_email saying the registration has been transferred.

* Send confirmation email from RegistrationTransferService

After updating the registration, send an email to the new account_email of the registration.

* Return :success_existing_user

When this service is called from a view and controller, we should return a status so we can display the correct response on the page.

* Add RegistrationTransfersController with views and routes

Create a new controller to let NCCC users transfer registrations in the browser. This will call the RegistrationTransferService.

* Use params to set @registraton in controller

The registration to transfer is determined based on the path. For example, /bo/transfer-registration/CBDU1234 should use the registration with that reg_identifier. We also use this to display information about the registration on the page so NCCC can confirm they are dealing with the correct registration before proceeding.

* Add some content to the transfer page

Explain how the transferring works, and display some information about the registration so the NCCC user can verify it's the right one.

* Add new :transfer_registration ability

Agency users (with and without refund) and agency super users should have this ability.

* Check authorisation of users attempting to transfer registrations

Now that we've set up our new :transfer_registration ability with CanCanCan, we can chuck this into the controller and stop unauthorised users from proceeding.

* Create RegistrationTransferForm model

We use the form objects pattern for this form, as while the BaseForm in the engine currently assumes we are dealing with a transient_registration, it should still work for registrations as well.

This model will be used to validate the form.

* Set up RegistrationTransfersController to use form object

Use the form object to validate the params submitted to the form.

This is based off of the AdminFormsController and is pretty similar, with the exception of being specific to a single form and using a registration instead of a transient_registration. Both of these are due a larger refactor to reduce duplication - however that's outside the scope of this PR and will be done separately.

* Transfer the registration if form is valid

When the form validates, call the RegistrationTransferService and transfer the registration to the account matching the provided email.

* Add form to view

Add the `email` and `confirm_email` fields to the page so the form params can be submitted using the browser.

* Add custom error messages

Leaving the fields blank should result in useful error messages, instead of just "can't be blank".

* Validate email fields properly

For the emails to be valid, they must be in the correct format for an email address and match each other. We already have custom validators to do this (originally written for the contact email page in the renewal journey) so we can reimplement them here.

* Rename `set_up_form`

This was aggravating Rubocop:

`Naming/AccessorMethodName: Do not prefix writer method names with set_. (https://github.com/rubocop-hq/ruby-style-guide#accessor_mutator_method_names)`

However this does mean it is less consistent with the other form object controllers.

* Add success page for valid form requests

When the user submits a valid transfer form, redirect them to a new page afterwards saying that the registration has been successfully transferred.

Additional steps are required next to handle non-successful outcomes from the transfer service.

* Validate form based on result from RegistrationTransferService

Call the service from the form model instead of the controller, then use the result to validate whether the form has been submitted successfully.

This implementation is more consistent with other forms which call services (like our address lookups or Companies House forms). It also means that we can treat a failure from the service the same way as any other validation error and can therefore display error messages consistently.

* Improve content of confirmation email

Make it clearer how to sign in and add some details about the registration to help the user verify what it is for.

* Rescue mailer errors in service

If the mailer encounters an error, this should not prevent the user from continuing and should instead let them know that the transfer has occurred (as this happens before the email is sent).

However, we do log the error and send it to Errbit.